:sparkles: Get affine transformation from GeoTIFF (#8)
* :heavy_plus_sign: Add geo

Geospatial primitives and algorithms!

* :sparkles: Get affine transformation from GeoTIFF

Implement transform method in CogReader struct to get affine transformation matrix from GeoTIFF via ModelPixelScaleTag and ModelTiepointTag. Added a unit test to check that the x/y pixel size and top left origin pixel coordinates is correct.

* :memo: Document transform method's internal properties

Explicitly mentioning the internal fields of the AffineTransform struct, specifically the pixel resolution, rotation and origin, and included a reference to the GeoTIFF standard. Also renamed origin_x/origin_y to x_origin/y_origin.

* :construction: Raise unimplemented when getting non-zero rotation values

Supposed to be parsed from the ModelTransformationTag, but haven't found a sample GeoTIFF online with non-zero rotation to implement this properly with a unit test, so marking as unimplemented for now.