:sparkles: CogReader ndarray method to decode TIFF into an ndarray Array (#10)
* :recycle: Turn image decoding into a Vec code into a method call

Making the decoder.read_image() code an impl method under the CogReader struct. Reworded the previous panic message into an unimplemented error with a more helpful message that dtypes other than float32 are not yet supported.

* :sparkles: Implement ndarray method in CogReader struct

Move the previous tiff-decoder -> Vec -> ndarray code to an impl method under CogReader. Was going to separate the as_vec and ndarray code, but got caught up with move semantics when trying to use the methods in pyo3, so just having an all-in-one ndarray method now. Added a unit test to check that an ndarray is properly returned.